Rename DelirCore.EffectPluginBase to DelirCore.PostEffectBase
