+++ src/delir-core/src/plugin-support/PostEffectBase.ts
