+++ src/delir-core/src/plugin-support/index.ts
@@ -1,5 +1,5 @@
 export {default as PluginBase} from './plugin-base'
-export {default as EffectPluginBase} from './effect-plugin-base'
+export {default as PostEffectBase} from './PostEffectBase'
 export {default as FSPluginLoader} from './FSPluginLoader'
 export {default as PluginRegistry} from './plugin-registry'
 export {default as TypeDescriptor} from './type-descriptor'
